fix(bin/recording): fix eyelinksdk segfault by init SDL
After a long debugging session with Phil, we determined from the
assembly that SDL_ListModes was segfaulting because one of the global
variables in SDL was not initialized. Although the eyelink sdk's
init_expt_grpahics() should theoretically initialize this using
SDL_Init(), it apparently wasn't happening correctly.

By initializing it ourselves before running get_display_into or
init_expt_graphics, we ensure that the variable is not null.