feat(bin/player): add placeholder for video player
My original experience piping the output of x264 into mpv for playing in
realtime was that mpv had too much latency to be useable in a realtime
setting, even when trying to play it with low latency. For example:

    ./x264 --preset fast --tune zerolatency --min-keyint 8000 --no-scenecut sample.y4m --qp 24 -o /dev/stdout | mpv --profile=low-tatency --untimed -fs -

This still seemed unusable. In an ideal world, we can find a simple
video player that is extremely low latency so we can just pipe to it and
not worry about writing our own. Otherwise, we will likely need to
implement our own player that is low latency. We can take some
simplifications (e.g., no audio).

This placeholder code is a first attempt at using ffmpeg and SDL2 to
play video. However, the current implementation is also too slow, as it
appears to sleep too long between frames (perhaps a bug on my part?).