refactor: add catch-up mechanism to correct for drift
When encoding, due to the variable time it takes to encode a frame, we
can "drift" of of the target frame duration if we don't exactly align
correctly with timing. Rather than allowing this drift to accumulate,
this adds a catch-up mechanism, where if the calculated frame number
based solely on input fps and time differs from where we actually are in
processing, we continue to advance frames until we catch up.