refactor(bin/latency): initial results of e2e timing
We currently are measuring huge numbers, like 74-84ms e2e delay. This
seems larger than I would expect as we are achieving ~51 FPS. Suggesting
this is taking multiple frames to propagate.

I need to double check this and find better way to get timing between
the points in the client and server.