feat(fvideo/latency): ugly implementation for e2e latency of 2stream
This is really ugly. I don't like that this is copy-pasted essentially
from the actual server implementations, which is an opportunity for
mistakes between implementations. This should be refactored so that the
actual servers contain a "dummy" mode where the input comes from dummy
buffers of video rather than reading from a file, but then uses the same
or very similar functions for the actual encoding.