fix(fvideo/client): fix pairing of gaze sample to display
To help with pipelining, we actually send a new gaze sample before we
display the frame from the previous gaze sample. That means that display
based on the most recent gaze sample is actually off by 1.

Even with this fix, however, there is still some noticable jitter in the
image. This is probably due to rounding errors as we go back and forth
between float and integer types to position the foreground.

Fixes: d630a4e