refactor(fvideo/client): use an RGBA texture to alpha blending
Uses ffmpeg to convert the yuv420p decoded frame into an ABGR frame.
This appears to be the required format to correctly update into an
RGBA8888 SDL2 texture.

This extra colorspace conversion (which does not have a special
accelerated version), drops the FPS from ~450fps to ~380fps on my
machine. A fairly significant performance hit (with f=10).

Using this new RGBA texture, we can blend the foreground with the
background. For this implementation, I just use a 2D Guassian
hand-tweaked until I felt it looked OK. In particular, I wanted a fairly
large section of the foreground to be maxed at 255, with a quick
drop-off on the edges. You can see what this contour looks like for f=10
by plotting the following in Wolfram Alpha:

    Plot3D[min(255, 512* e^(-(((x - 80)^2 + (y - 80)^2) / (2 * (160/3.5)^2)))), {x, 0, 160}, {y, 0, 160}]

The additional operations to compute these alpha values does not
introduce any significant performance hit compared to the colorspace
conversion.