fix(fvideo/twostreamserver): fix incorrect crop at top and left edges
This is really a hacky fix. Rather than filling with 0s or some other
safer mechanism, I'm allowing the crop operation to reach beyond the
bounds of the image (which was already happening at the bottom and right
edges of the frame).

To the user, this is fine since that garbage data is hidden when
displayed. But, this feels unpleasant and should be revisited.