feat(fvideo): allow artificial delay to be added to gaze samples
This is not as simple as adding a sleep() somewhere. We still want all
the smooth gaze data, just delayed by a fixed offset of time. So, we
need to introduce a queue to store these gaze samples while they queue
up during the delay.

Unfortunately, this hurts our e2e latency significantly (dropping us from
~18ms to like 22-24ms). We will need to optimize this.