perf(fvideo/client): avoid unaccelerated colorspace conversion
When introducing the RGBA texture for alpha blending, we introduced a
colorspace conversion that ffmpeg did not accelerate (yuv420p to abgr).
This unaccelerated conversion can be easily replaced with an accelerated
version (yuv420p to rgba) by using the correct pixel format for the
foreground texture. Somewhat counterintuitively, this is ABGR8888.

It is unclear to me why these appear to be inverses of each other, but
it is functional and the performance drop incurred previously is
successfully avoided.

Fixes: 222926cebf0dac4a94e8936237c4c197b4c2c250