refactor: only send new foreground if delta > thresh
The goal of this is to start addressing the low-hanging fruit for
reducing the foreground size. One of those is not sending repetitive
frames. For example, in a 30fps video, there is 33 ms between frames.
Within those 33ms, the eye may or may not move very significantly.

Previously, the server/client would simply stream foreground frames as
fast as possible, even if the eye was stationary and the frames were
completely redundant. This now only send new foreground frames when
either the frame advances, or if the gaze position changes above some
threshold.

This optimization seems to have very little effect (e.g., dropping from
~17MB foreground to ~15MB foreground in the best case (where the gaze
does not move whatsoever). Much less frames total are transmitted
though.