refactor: assume vid resolution <= display resolution
This allows us to compute the video pixel position directly for the gaze
and will allow us to show black bars around content that is smaller than
the display resolution.

Previously, we would implicitly scale the output to match the display
resolution. This change eliminates the need for that scaling step.

BREAKING CHANGE: This fundamentally changes the assumption that the code
is built on. Rather than being able to take any resolution of video and
fit it to any resolution display, we now require that the source video
resolution be less than or equal to the display resolution in both width
and height.