fix(fvideo): fix the trembling of the foreground
The root cause is that when cropping, the server might make an
adjustment of 1 pixel to make it the top and left values multiples of 2.
This 1 pixel adjustment of the server could put the display 1 pixel off
as well, but only when the values weren't multiples of 2, resulting in
the trembling.

Combined with 13903c4bd56118a46e50002e2332787538dd8435, we can tell the
client the exact gaze position used.