fix(eyelink-rs): fix screen_write_prescale to support 4k
Previously, the x-axis value of the gaze from the eyelink was being
maxed at 3270, less than the full 3840 width of the display. This has to
do with the maximum value being 3270 -- This is due to the value of the
"screen_write_prescale" setting on the Host PC

Internally, the EyeLink 1000 software sets an upper bounds value of
32700 when storing data. If the prescale is set to 10, then it will only
be able to report data up to 3270. You can use a lower prescale value
like 4 (this will report values up to 8100) or lower.

This commit adjusts the value to 6.

See: https://www.sr-support.com/thread-7414.html