feat(fvideo): track simple gaze statistics
Logs the min/max x,y pixel values the gaze reached, as well as the total
"distance" in x, y values covered during a video. This distance is
defined as just the sum of all of the movements between gazes, and gives
a rough idea of how much a viewer's gaze moved around over the course of
the video.