data: add 300 trials minlatency.csv
This is using my desktop and the LG 27GN95B-B display (3840x2160@144Hz).
However, this is also with a big grain of salt in that it's barely
better than the latency numbers that actually include encode/decode. I
believe this is due to inefficiency in how I've implemented the client's
display_white() function.

In the C version we used previously, this would've been switching
between pre-allocated textures and I would expect that to be noticeably
faster. So, I think the true min latency numbers are less than what is
shown here.