refactor(fvideo/client): simplify artificial delay code
Though, I'm not sure this is great as is. I can definitely perceive that
larger delays do in fact look like larger delays. But, it also seems
that switching from a delay of 0 to a delay of 1 has more visual impact
than I would expect.

I suspect that comparison of Instant/Duration is slow.