fix(fvideo): drain artificial delay queue properly
As implemented before, we only release a single sample per gaze_sample()
call. This is bad because there is head-of-line blocking.

At each call, we should instead return the latest eligible gaze sample
(per the delay setting), such that the next sample in the queue is
always the first one that did not have sufficient time elapsed.

In the delay=0 case, this queue should always be empty.