fix(fvideo): blank screen and minimize window right after
This prevents the situation where a user is left staring at a mostly
poor-quality frame at the end of an experiment.

Ideally, I could just clear() and that's it, but I found that SDL would
still flash a frame as it shutdown (it looked like the first frame of
the video). So, the minimize is a hacky workaround. If it's minimized,
at least the user won't see that random frame flash.